feat: add DeviceItem into the Onboarding (reuse from ViewOnlyPromo)
